Finish exercise 4.21*: blogilistan laajennus, step10

+++ osa4/controllers/blogs.js
@@ -39,6 +39,16 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!(blog.user.toString() === decodedToken.id)) {
+      return response.status(401).json({ error: 'user is not authorized to delete this blog' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch(exception) {
